Split PegtlTests.cpp into 3 grammar variants to speeed up build
